requirements.txt
Flask==3.0.3
protobuf==5.27.2
python-dotenv==1.0.1
sentence_transformers==3.0.1
